+++ src/my_first_adventure_game/game/entities/__init__.py
@@ -1,7 +1,17 @@
+from my_first_adventure_game.game.entities.caretaker_movement import (
+    caretaker_sidestep_target,
+)
 from my_first_adventure_game.game.entities.enemy import Enemy
 from my_first_adventure_game.game.entities.npc import NPC
 from my_first_adventure_game.game.entities.npc_movement import move_npc_towards
 from my_first_adventure_game.game.entities.player import Player
 from my_first_adventure_game.game.entities.wall_stain import WallStain
 
-__all__ = ["Enemy", "NPC", "Player", "WallStain", "move_npc_towards"]
+__all__ = [
+    "Enemy",
+    "NPC",
+    "Player",
+    "WallStain",
+    "caretaker_sidestep_target",
+    "move_npc_towards",
+]